Refactor dump command to use prompt
