+++ src/commands/prompts/dump/index.js
@@ -1,0 +1,1 @@
+exports.sendFile = require('./sendFile.js')
